EnemyController Modularisation and Enemy Type 2
General Enemy behaviour placed in abstract EnemyController class which is inherited by EnemyController1 (enemy which does not attack player) and EnemyController2 (enemy which will attack attack player, to-be-implemented). Enemy 2 sprite animation setup. EnemyController2 has slight modifications from EnemyController1 for damaging. For EnemyController1, damage is applied as soon as the attack animation of the wolf plays. For EnemyController2, damage is only applied when the 8th of 18 frames of the skeleton attack spritesheet is played (when its axe is swung).